requirements.txt
fastapi==0.115.2
starlette==0.40.0
pydantic==2.9.2
uvicorn==0.30.6
sqlalchemy==2.0.36
python-dotenv==1.0.1
websockets==12.0
